requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.6
fonttools==4.59.2
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==6.0.1
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
boto3==1.35.73
django-storages==1.14.4
djangorestframework==3.15.2
